# Add sha3_256 and sha3_512 to preferred algorithms